pyproject.toml
[project]
name = "blendbridge"
version = "0.1.0"
description = "MCP server for controlling Blender — create low-poly 3D models with AI"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "pillow",
    "pyyaml",
]

[project.scripts]
blendbridge = "mcp_server:mcp.run"
